Implemented heat bar. Purged shields and replaced with a dissipation bar that constantly lowers and is raised by firing and using abilities. When bar goes over limit, player cannot fire or use abilities until the bar has decreased to 75%. Abilities present in this branch have been modified to utilize heat.

+++ Assets/src/Abilities/Utils/AbilityUtils.cs
@@ -9,7 +9,7 @@ public class BaseAbility: MonoBehaviour{
 	protected ShipAction Ship;
 	protected ShipMovement ShipMove;
 	public bool Executing {get; set;}
-	public int Cost {get; set;}
+	public float Cost {get; set;}
 	public int Level = 1;
 	public Object Resource;
